Adjust risk indexing to use 11 bins, including 0-0.5 and 0.95-1
